Safari-609.1.11.1/LayoutTests/js/exception-for-nonobject-expected.txt

Test for correct handling of exceptions from instanceof and 'new' expressions

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS new {}.undefined threw exception TypeError: undefined is not a constructor (evaluating 'new {}.undefined').
PASS 1 instanceof {}.undefined threw exception TypeError: Right hand side of instanceof is not an object.
PASS successfullyParsed is true

TEST COMPLETE
